acc: Fix drop_accounting() (do not be a NOOP!)
The bug here was that `drop_accounting()` was building its bitmask as:

    flag_mask = types * flags;

... whereas it should have done the same as in `do_accounting()`:

    flag_mask = types + types * flags;
                -------
		^ the actual acc types to reset were missing!

(cherry picked from commit 211a63c9b2ccac346dfc57bca2ede71d286e9989)

+++ modules/acc/acc_logic.h
@@ -33,11 +33,11 @@
 #include "../tm/t_hooks.h"
 #include "../dialog/dlg_cb.h"
 
-#define DO_ACC_NONE (0)
-#define DO_ACC_LOG  (1<<(0*8))
-#define DO_ACC_AAA  (1<<(1*8))
-#define DO_ACC_DB   (1<<(2*8))
-#define DO_ACC_EVI  ((unsigned long long)1<<(4*8))
+#define DO_ACC_NONE (0ULL)
+#define DO_ACC_LOG  (1ULL<<(0*8))
+#define DO_ACC_AAA  (1ULL<<(1*8))
+#define DO_ACC_DB   (1ULL<<(2*8))
+#define DO_ACC_EVI  (1ULL<<(4*8))
 #define DO_ACC_ERR  ((unsigned long long)-1)
 
 #define DO_ACC        (1<<0) /* generic accouting flag - internal only */
